backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pytest
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
python-multipart
email-validator
dnspython
pydantic
bcrypt<4

